refactor: move the generic date formatters out of the table domain
`formatDateValue`/`preciseDateTime` (and their `FormatOptions`/`DateConfig`
types) are generic Intl date formatting, but lived in `table/format.ts` — so
`format/value.ts` and `i18n/date-time.tsx` reached *up* into the table leaf
(format→table, i18n→table inversions). Move them to `format/date-time.ts`;
`table/format` now depends on `format/` (correct direction), and format/i18n no
longer import table at all. Table-specific `formatCell`/`resolveLink` stay put.
