+++ resources/js/table/format.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
 import type { TableColumn, TableRow } from "./types";
-import { formatCell, formatDateValue, preciseDateTime, resolveLink } from "./format";
+import { formatCell, resolveLink } from "./format";
 
 const dateColumn = {
   key: "created",
@@ -28,22 +28,6 @@ describe("formatCell date rendering", () => {
   });
 });
 
-describe("preciseDateTime", () => {
-  it("includes the IANA zone id and a year", () => {
-    const text = preciseDateTime("2026-06-18T00:30:00Z", {
-      locale: "en-GB",
-      timeZone: "Europe/Berlin",
-    });
-
-    expect(text).toContain("2026");
-    expect(text).toContain("Europe/Berlin");
-  });
-
-  it("returns an empty string for an invalid value", () => {
-    expect(preciseDateTime("not-a-date", { timeZone: "UTC" })).toBe("");
-  });
-});
-
 describe("formatCell primitives", () => {
   it("returns an empty string for null and undefined", () => {
     expect(formatCell(null)).toBe("");
@@ -58,36 +42,6 @@ describe("formatCell primitives", () => {
 
   it("JSON-encodes non-primitive values", () => {
     expect(formatCell({ a: 1 })).toBe('{"a":1}');
-  });
-});
-
-describe("formatDateValue", () => {
-  it("returns the raw value when the date cannot be parsed", () => {
-    expect(formatDateValue("not-a-date", { dateStyle: "medium", timeStyle: null })).toBe(
-      "not-a-date",
-    );
-  });
-
-  it("applies only the styles that are configured", () => {
-    const dateOnly = formatDateValue(
-      "2026-06-18T00:30:00Z",
-      { dateStyle: "short", timeStyle: null },
-      {
-        locale: "en-GB",
-        timeZone: "UTC",
-      },
-    );
-    const timeOnly = formatDateValue(
-      "2026-06-18T00:30:00Z",
-      { dateStyle: null, timeStyle: "short" },
-      {
-        locale: "en-GB",
-        timeZone: "UTC",
-      },
-    );
-
-    expect(dateOnly).toContain("2026");
-    expect(timeOnly).toContain("00:30");
   });
 });
