+++ resources/js/table/format.ts
@@ -1,16 +1,5 @@
+import { type FormatOptions, formatDateValue } from "@lattice-php/lattice/format/date-time";
 import type { ColumnPropsOf, TableColumn, TableRow } from "./types";
-
-export type FormatOptions = {
-  locale?: string;
-  timeZone?: string;
-};
-
-export type DateConfig = {
-  dateStyle: string | null;
-  timeStyle: string | null;
-  month?: string | null;
-  year?: string | null;
-};
 
 export function formatCell(value: unknown, column?: TableColumn, options?: FormatOptions): string {
   if (value === null || value === undefined) {
@@ -28,55 +17,6 @@ export function formatCell(value: unknown, column?: TableColumn, options?: Forma
   }
 
   return JSON.stringify(value);
-}
-
-export function formatDateValue(value: unknown, date: DateConfig, options?: FormatOptions): string {
-  const parsed = new Date(String(value));
-
-  if (Number.isNaN(parsed.getTime())) {
-    return String(value ?? "");
-  }
-
-  const intl: Intl.DateTimeFormatOptions = { timeZone: options?.timeZone };
-
-  if (date.dateStyle) {
-    intl.dateStyle = date.dateStyle as Intl.DateTimeFormatOptions["dateStyle"];
-  }
-
-  if (date.timeStyle) {
-    intl.timeStyle = date.timeStyle as Intl.DateTimeFormatOptions["timeStyle"];
-  }
-
-  if (date.month) {
-    intl.month = date.month as Intl.DateTimeFormatOptions["month"];
-  }
-
-  if (date.year) {
-    intl.year = date.year as Intl.DateTimeFormatOptions["year"];
-  }
-
-  return new Intl.DateTimeFormat(options?.locale, intl).format(parsed);
-}
-
-export function preciseDateTime(value: unknown, options?: FormatOptions): string {
-  const date = new Date(String(value));
-
-  if (Number.isNaN(date.getTime())) {
-    return "";
-  }
-
-  const formatted = new Intl.DateTimeFormat(options?.locale, {
-    year: "numeric",
-    month: "short",
-    day: "numeric",
-    hour: "2-digit",
-    minute: "2-digit",
-    second: "2-digit",
-    timeZone: options?.timeZone,
-    timeZoneName: "short",
-  }).format(date);
-
-  return options?.timeZone ? `${formatted} (${options.timeZone})` : formatted;
 }
 
 export function resolveLink(column: TableColumn, row: TableRow, value: unknown): string | null {
